Clean up ViewModels and Login - Some ViewModels were in the Models package instead of ViewModels.  Finished moving them over. - Failed logins would result in a stack trace due to referencing an old LoginViewModel class.  Fixed.
